% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kquarto.r.inrae-package.R
\docType{package}
\name{kquarto.r.inrae-package}
\alias{kquarto.r.inrae}
\alias{kquarto.r.inrae-package}
\title{kquarto.r.inrae: 'Quarto' Book Templates Following the INRAE Graphic Charter}
\description{
Ships an 'INRAE' template for the 'Quarto' books created with 'kquarto.r' (colours, fonts, logo, title page and bibliography) and makes it the default template of the book creation helpers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbosirany/kquarto.r.inrae}
  \item \url{https://kbosirany.github.io/kquarto.r.inrae/}
  \item Report bugs at \url{https://github.com/kbosirany/kquarto.r.inrae/issues}
}

}
\author{
\strong{Maintainer}: Kevin Bosirany Orlando \email{kevinbosirany@gmail.com} (\href{https://orcid.org/0009-0009-2784-3108}{ORCID})

}
\keyword{internal}
