% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_book.R
\name{list_templates_book}
\alias{list_templates_book}
\title{List the available book templates, INRAE template included}
\usage{
list_templates_book(packages = "kquarto.r.inrae")
}
\arguments{
\item{packages}{Names of the packages whose templates are listed too.}
}
\value{
A data frame with the template \code{name}, \code{source} and \code{path}.
}
\description{
Same as \code{\link[kquarto.r:list_templates_book]{kquarto.r::list_templates_book()}}, also listing the templates of
this package.
}
\examples{
list_templates_book()
}
