% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{render_book}
\alias{create_book_chapter}
\alias{list_books}
\alias{remove_template_book}
\alias{template_dir}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{kquarto.r}{\code{\link[kquarto.r]{create_book_chapter}}, \code{\link[kquarto.r]{list_books}}, \code{\link[kquarto.r]{remove_template_book}}, \code{\link[kquarto.r]{render_book}}, \code{\link[kquarto.r]{template_dir}}}
}}

